Move rrdlabels api to separate file.
Also, move localhost functions that loads
labels outside of database/ and into daemon/
