+++ daemon/commands.c
@@ -218,7 +218,7 @@ static cmd_status_t cmd_reload_labels_execute(char *args, char **message)
 {
     (void)args;
     netdata_log_info("COMMAND: reloading host labels.");
-    reload_host_labels();
+    localhost_load_labels();
 
     BUFFER *wb = buffer_create(10, NULL);
     rrdlabels_log_to_buffer(rrdb.localhost->rrdlabels, wb);
